delete sql

+++ Nodes/node/app/model/select.js
@@ -59,7 +59,7 @@ exports.selectGameList = function(response){
 	con.connect(function(err) {
 	  if (err) throw err;
 	  con.query(
-		"SELECT game_name, games_review FROM games join games_list on games_list.game_id = games.game_id where user_id = ?"[userID]
+		"SELECT game_name, games_review FROM games join games_list on games_list.game_id = games.game_id where user_id = ?"[user_Id]
 		, function (err, result, fields) { // change the query to match your needs and setup
 		if (err) throw err;
 		console.log("From model: " + result);
@@ -84,7 +84,33 @@ exports.deleteFromGameList = function(response){
 	con.connect(function(err) {
 	  if (err) throw err;
 	  con.query(
-		"DELETE FROM games_list where game_id = ?"[game_Id]
+		"DELETE FROM games_list where game_id = ? and user_id = ?"[game_Id, user_Id]
+		, function (err, result, fields) { // change the query to match your needs and setup
+		if (err) throw err;
+		console.log("From model: " + result);
+		response(result);		
+		
+	  });
+	});
+}
+
+
+exports.updateGameList = function(response){
+
+	var mysql = require('mysql2');
+
+	var con = mysql.createConnection({ // change these details to match your installation
+	  host: "localhost",
+	  user: "root",
+	  password: "1234", 
+	  database: "sys",
+	  port:3306
+	});
+
+	con.connect(function(err) {
+	  if (err) throw err;
+	  con.query(
+		"UPDATE games_list set games_review = ? where game_id = ? and user_id = ?"[game_Id, user_Id]
 		, function (err, result, fields) { // change the query to match your needs and setup
 		if (err) throw err;
 		console.log("From model: " + result);
